Improve code quality
The static method parse in Parser class contained all the logic for
translating the input string from the user into usable arguments for the
Command classes to use. However, this is a poor use of SLAP principal
and makes for poor code readability as the method is too long.

Thus to improve on the code quality and readability, the logic used to
translate the input string into usable formats for each Command, was
extracted and put into the Command classes, for the Command classes to
define the logic to change the input into their respective formats.